fix(test): cast `-$3` → `-$3::integer` in MSP refund SQL
Postgres planner kann `-$N` ohne expliziten Cast als "unary minus on
unknown" nicht eindeutig auflösen — "operator is not unique: - unknown".
postgres-js hat das anders inferred, postgres 18 (oder die neue
type-passing-Path) ist strikter.

Diagnose: refund-handler WIRD aufgerufen, SQL throws,
event-dispatcher catched (skipApplyErrors), cursor advanced, last_error
gespeichert. Effekt: refund wird nie applied, netCents bleibt
shipment-only.

Fix: `-$3` → `-$3::integer` an beiden Stellen im refund-handler-SQL.

Stand: 3720 pass / 0 fail (von 3720 total).

+++ packages/framework/src/pipeline/__tests__/multi-stream-projection.integration.test.ts
@@ -73,8 +73,10 @@ const mspFeature = defineFeature("msptest", (r) => {
       },
       [refundIssued.name]: async (event, tx) => {
         const p = event.payload as { customer: string; cents: number };
+        // -$3 ohne cast wird vom planner als "unary minus on unknown type"
+        // mehrdeutig — wir casten $3 explizit auf integer.
         await asRawClient(tx).unsafe(
-          `INSERT INTO "read_msp_customer_balance" (customer, tenant_id, shipments, refunds, net_cents) VALUES ($1::uuid, $2::uuid, 0, 1, -$3) ON CONFLICT (customer) DO UPDATE SET refunds = read_msp_customer_balance.refunds + 1, net_cents = read_msp_customer_balance.net_cents - $3`,
+          `INSERT INTO "read_msp_customer_balance" (customer, tenant_id, shipments, refunds, net_cents) VALUES ($1::uuid, $2::uuid, 0, 1, -$3::integer) ON CONFLICT (customer) DO UPDATE SET refunds = read_msp_customer_balance.refunds + 1, net_cents = read_msp_customer_balance.net_cents - $3::integer`,
           [p.customer, event.tenantId, p.cents],
         );
       },
